refactor: Update MypageController.java
