--- src/main/java/com/kurlabo/backend/converter/StringRevisor.java
@@ -1,10 +1,0 @@
-package com.kurlabo.backend.converter;
-
-public class StringRevisor {
-    public String reviseBackSlash(String origin){
-        return origin.replace("\\", "");
-    }
-    public String reviseDoubleQuotes(String origin){
-        return origin.replace("\"", "");
-    }
-}
