--- src/main/java/com/kurlabo/backend/dto/member/MemberResponseDto.java
@@ -1,19 +1,0 @@
-package com.kurlabo.backend.dto.member;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MemberResponseDto {
-        @NotNull
-        @NotEmpty
-        private String message;
-}
