--- src/main/java/com/kurlabo/backend/dto/mypage/GetMyInfoResponseDto.java
@@ -1,18 +1,0 @@
-package com.kurlabo.backend.dto.mypage;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class GetMyInfoResponseDto {
-    @NotNull @NotEmpty
-    private String uid;
-}
