--- src/main/java/com/kurlabo/backend/dto/order/OrderListDto.java
@@ -1,15 +1,0 @@
-package com.kurlabo.backend.dto.order;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class OrderListDto {
-    private Long product_id;
-    private int cnt;
-}
