--- src/main/java/com/kurlabo/backend/dto/order/OrderSheetRequestDto.java
@@ -1,14 +1,0 @@
-package com.kurlabo.backend.dto.order;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class OrderSheetRequestDto {
-    private Long member_id;
-}
