--- src/main/java/com/kurlabo/backend/dto/testdto/CartTestDto.java
@@ -1,15 +1,0 @@
-package com.kurlabo.backend.dto.testdto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CartTestDto {
-    List<ProductInfoTestDto> productInfoTestDto;
-    UserInfoTestDto userInfoTestDto;
-}
