--- src/main/java/com/kurlabo/backend/dto/testdto/FindIdTestDto.java
@@ -1,13 +1,0 @@
-package com.kurlabo.backend.dto.testdto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class FindIdTestDto {
-    private String name;
-    private String email;
-}
