--- src/main/java/com/kurlabo/backend/dto/testdto/TestInfoDto.java
@@ -1,14 +1,0 @@
-package com.kurlabo.backend.dto.testdto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class TestInfoDto {
-    private String token;
-    private Long id;
-    private String uid;
-}
