--- src/main/java/com/kurlabo/backend/dto/testdto/WishListTestDto.java
@@ -1,14 +1,0 @@
-package com.kurlabo.backend.dto.testdto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class WishListTestDto {
-    private String name;
-    private int discounted_price;
-    private String list_image_url;
-}
